Feature refactoring
fixed some of the issues occurring with the cart items not being shown when deleted in real time. Also fixed the animation function to work everytime the button is clicked. Also added a if statement to our mainheader.svelte file to only show the superscript number when there are items in the cart.